requirements.txt
PyQt5==5.15.11
pillow==11.0.0
